[resolve-conflict-IC-1320] salary
