+++ blockchain/committeestate/beacon_committee_state_v1.go
@@ -433,45 +433,6 @@ func (b *BeaconCommitteeStateV1) processAutoStakingChange(committeeChange *Commi
 	return nil
 }
 
-//SplitReward ...
-func (b *BeaconCommitteeStateV1) SplitReward(
-	env *SplitRewardEnvironment) (
-	map[common.Hash]uint64, map[common.Hash]uint64,
-	map[common.Hash]uint64, map[common.Hash]uint64, error,
-) {
-
-	devPercent := uint64(env.DAOPercent)
-	allCoinTotalReward := env.TotalReward
-	rewardForBeacon := map[common.Hash]uint64{}
-	rewardForShard := map[common.Hash]uint64{}
-	rewardForIncDAO := map[common.Hash]uint64{}
-	rewardForCustodian := map[common.Hash]uint64{}
-
-	if len(allCoinTotalReward) == 0 {
-		Logger.log.Info("Beacon Height %+v, 😭 found NO reward", env.BeaconHeight)
-		return rewardForBeacon, rewardForShard, rewardForIncDAO, rewardForCustodian, nil
-	}
-
-	for key, totalReward := range allCoinTotalReward {
-		rewardForBeacon[key] += 2 * ((100 - devPercent) * totalReward) / ((uint64(env.ActiveShards) + 2) * 100)
-		totalRewardForDAOAndCustodians := uint64(devPercent) * totalReward / uint64(100)
-
-		Logger.log.Infof("[test-salary] totalRewardForDAOAndCustodians tokenID %v - %v\n",
-			key.String(), totalRewardForDAOAndCustodians)
-
-		if env.IsSplitRewardForCustodian {
-			rewardForCustodian[key] += uint64(env.PercentCustodianReward) * totalRewardForDAOAndCustodians / uint64(100)
-			rewardForIncDAO[key] += totalRewardForDAOAndCustodians - rewardForCustodian[key]
-		} else {
-			rewardForIncDAO[key] += totalRewardForDAOAndCustodians
-		}
-
-		rewardForShard[key] = totalReward - (rewardForBeacon[key] + totalRewardForDAOAndCustodians)
-	}
-
-	return rewardForBeacon, rewardForShard, rewardForIncDAO, rewardForCustodian, nil
-}
-
 func (b *BeaconCommitteeStateV1) GenerateAssignInstructions(env *BeaconCommitteeStateEnvironment) []*instruction.AssignInstruction {
 	candidates := make([]string, len(b.currentEpochShardCandidate))
 	copy(candidates, b.currentEpochShardCandidate)
