+++ blockchain/committeestate/beacon_committee_state_v2.go
@@ -238,78 +238,6 @@ func (b *BeaconCommitteeStateV2) getDataForUpgrading(env *BeaconCommitteeStateEn
 		autoStake, rewardReceiver, stakingTx
 }
 
-//SplitReward ...
-func (b *BeaconCommitteeStateV2) SplitReward(
-	env *SplitRewardEnvironment) (
-	map[common.Hash]uint64, map[common.Hash]uint64,
-	map[common.Hash]uint64, map[common.Hash]uint64, error,
-) {
-	// @NOTICE: No use split rule reward v2
-	/*
-		devPercent := uint64(env.DAOPercent)
-		allCoinTotalReward := env.TotalReward
-		rewardForBeacon := map[common.Hash]uint64{}
-		rewardForShard := map[common.Hash]uint64{}
-		rewardForIncDAO := map[common.Hash]uint64{}
-		rewardForCustodian := map[common.Hash]uint64{}
-		lenBeaconCommittees := uint64(len(b.getBeaconCommittee()))
-		lenShardCommittees := uint64(len(b.getShardCommittee()[env.ShardID]))
-		if len(allCoinTotalReward) == 0 {
-			Logger.log.Info("Beacon Height %+v, 😭 found NO reward", env.BeaconHeight)
-			return rewardForBeacon, rewardForShard, rewardForIncDAO, rewardForCustodian, nil
-		}
-		for key, totalReward := range allCoinTotalReward {
-			totalRewardForDAOAndCustodians := devPercent * totalReward / 100
-			totalRewardForShardAndBeaconValidators := totalReward - totalRewardForDAOAndCustodians
-			shardWeight := float64(lenShardCommittees)
-			beaconWeight := 2 * float64(lenBeaconCommittees) / float64(len(b.shardCommittee))
-			totalValidatorWeight := shardWeight + beaconWeight
-			rewardForShard[key] = uint64(shardWeight * float64(totalRewardForShardAndBeaconValidators) / totalValidatorWeight)
-			Logger.log.Infof("[test-salary] totalRewardForDAOAndCustodians tokenID %v - %v\n",
-				key.String(), totalRewardForDAOAndCustodians)
-			if env.IsSplitRewardForCustodian {
-				rewardForCustodian[key] += env.PercentCustodianReward * totalRewardForDAOAndCustodians / 100
-				rewardForIncDAO[key] += totalRewardForDAOAndCustodians - rewardForCustodian[key]
-			} else {
-				rewardForIncDAO[key] += totalRewardForDAOAndCustodians
-			}
-			rewardForBeacon[key] += totalReward - (rewardForShard[key] + totalRewardForDAOAndCustodians)
-		}
-		return rewardForBeacon, rewardForShard, rewardForIncDAO, rewardForCustodian, nil
-	**/
-
-	devPercent := uint64(env.DAOPercent)
-	allCoinTotalReward := env.TotalReward
-	rewardForBeacon := map[common.Hash]uint64{}
-	rewardForShard := map[common.Hash]uint64{}
-	rewardForIncDAO := map[common.Hash]uint64{}
-	rewardForCustodian := map[common.Hash]uint64{}
-
-	if len(allCoinTotalReward) == 0 {
-		Logger.log.Info("Beacon Height %+v, 😭 found NO reward", env.BeaconHeight)
-		return rewardForBeacon, rewardForShard, rewardForIncDAO, rewardForCustodian, nil
-	}
-
-	for key, totalReward := range allCoinTotalReward {
-		rewardForBeacon[key] += 2 * ((100 - devPercent) * totalReward) / ((uint64(env.ActiveShards) + 2) * 100)
-		totalRewardForDAOAndCustodians := uint64(devPercent) * totalReward / uint64(100)
-
-		Logger.log.Infof("[test-salary] totalRewardForDAOAndCustodians tokenID %v - %v\n",
-			key.String(), totalRewardForDAOAndCustodians)
-
-		if env.IsSplitRewardForCustodian {
-			rewardForCustodian[key] += uint64(env.PercentCustodianReward) * totalRewardForDAOAndCustodians / uint64(100)
-			rewardForIncDAO[key] += totalRewardForDAOAndCustodians - rewardForCustodian[key]
-		} else {
-			rewardForIncDAO[key] += totalRewardForDAOAndCustodians
-		}
-
-		rewardForShard[key] = totalReward - (rewardForBeacon[key] + totalRewardForDAOAndCustodians)
-	}
-
-	return rewardForBeacon, rewardForShard, rewardForIncDAO, rewardForCustodian, nil
-}
-
 func (b *beaconCommitteeStateSlashingBase) addData(env *BeaconCommitteeStateEnvironment) {
 	env.newUnassignedCommonPool = common.DeepCopyString(b.shardCommonPool[b.numberOfAssignedCandidates:])
 	env.newAllSubstituteCommittees, _ = b.getAllSubstituteCommittees()
