+++ blockchain/committeestate/beacon_committee_state_v3_test.go
@@ -5003,352 +5003,3 @@ func TestBeaconCommitteeStateV3_assignBackToSubstituteList(t *testing.T) {
 		})
 	}
 }
-
-func TestBeaconCommitteeStateV3_SplitReward(t *testing.T) {
-
-	initTestParams()
-
-	type fields struct {
-		beaconCommitteeStateSlashingBase beaconCommitteeStateSlashingBase
-		syncPool                         map[byte][]string
-	}
-	type args struct {
-		env *SplitRewardEnvironment
-	}
-	tests := []struct {
-		name    string
-		fields  fields
-		args    args
-		want    map[common.Hash]uint64
-		want1   map[common.Hash]uint64
-		want2   map[common.Hash]uint64
-		want3   map[common.Hash]uint64
-		wantErr bool
-	}{
-		{
-			name: "Year 1",
-			fields: fields{
-				beaconCommitteeStateSlashingBase: beaconCommitteeStateSlashingBase{
-					assignRule: NewAssignRuleV2(),
-					beaconCommitteeStateBase: beaconCommitteeStateBase{
-						beaconCommittee: []string{
-							key0, key, key2, key3, key, key2, key3,
-						},
-						shardCommittee: map[byte][]string{
-							0: []string{},
-							1: []string{
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3,
-							},
-							2: []string{},
-							3: []string{},
-							4: []string{},
-							5: []string{},
-							6: []string{},
-							7: []string{},
-						},
-						mu: &sync.RWMutex{},
-					},
-				},
-			},
-			args: args{
-				env: &SplitRewardEnvironment{
-					DAOPercent:                10,
-					ShardID:                   1,
-					IsSplitRewardForCustodian: false,
-					PercentCustodianReward:    0,
-					TotalReward: map[common.Hash]uint64{
-						common.PRVCoinID: 546998,
-					},
-					MaxSubsetCommittees: 2,
-					SubsetID:            1,
-				},
-			},
-			want: map[common.Hash]uint64{
-				common.PRVCoinID: 13104,
-			},
-			want1: map[common.Hash]uint64{
-				common.PRVCoinID: 479195,
-			},
-			want2: map[common.Hash]uint64{
-				common.PRVCoinID: 54699,
-			},
-			want3:   map[common.Hash]uint64{},
-			wantErr: false,
-		},
-		{
-			name: "Year 2, Subset 0",
-			fields: fields{
-				beaconCommitteeStateSlashingBase: beaconCommitteeStateSlashingBase{
-					assignRule: NewAssignRuleV2(),
-					beaconCommitteeStateBase: beaconCommitteeStateBase{
-						beaconCommittee: []string{
-							key0, key, key2, key3, key, key2, key3,
-						},
-						shardCommittee: map[byte][]string{
-							0: []string{},
-							1: []string{
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3,
-							},
-							2: []string{},
-							3: []string{},
-							4: []string{},
-							5: []string{},
-							6: []string{},
-							7: []string{},
-						},
-						mu: &sync.RWMutex{},
-					},
-				},
-			},
-			args: args{
-				env: &SplitRewardEnvironment{
-					DAOPercent:                9,
-					ShardID:                   1,
-					IsSplitRewardForCustodian: false,
-					PercentCustodianReward:    0,
-					TotalReward: map[common.Hash]uint64{
-						common.PRVCoinID: 497768,
-					},
-					MaxSubsetCommittees: 2,
-					SubsetID:            0,
-				},
-			},
-			want: map[common.Hash]uint64{
-				common.PRVCoinID: 12057,
-			},
-			want1: map[common.Hash]uint64{
-				common.PRVCoinID: 440912,
-			},
-			want2: map[common.Hash]uint64{
-				common.PRVCoinID: 44799,
-			},
-			want3:   map[common.Hash]uint64{},
-			wantErr: false,
-		},
-		{
-			name: "Year 2, Subset 1",
-			fields: fields{
-				beaconCommitteeStateSlashingBase: beaconCommitteeStateSlashingBase{
-					assignRule: NewAssignRuleV2(),
-					beaconCommitteeStateBase: beaconCommitteeStateBase{
-						beaconCommittee: []string{
-							key0, key, key2, key3, key, key2, key3,
-						},
-						shardCommittee: map[byte][]string{
-							0: []string{},
-							1: []string{
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3,
-							},
-							2: []string{},
-							3: []string{},
-							4: []string{},
-							5: []string{},
-							6: []string{},
-							7: []string{},
-						},
-						mu: &sync.RWMutex{},
-					},
-				},
-			},
-			args: args{
-				env: &SplitRewardEnvironment{
-					DAOPercent:                9,
-					ShardID:                   1,
-					IsSplitRewardForCustodian: false,
-					PercentCustodianReward:    0,
-					TotalReward: map[common.Hash]uint64{
-						common.PRVCoinID: 497768,
-					},
-					MaxSubsetCommittees: 2,
-					SubsetID:            1,
-				},
-			},
-			want: map[common.Hash]uint64{
-				common.PRVCoinID: 12057,
-			},
-			want1: map[common.Hash]uint64{
-				common.PRVCoinID: 440912,
-			},
-			want2: map[common.Hash]uint64{
-				common.PRVCoinID: 44799,
-			},
-			want3:   map[common.Hash]uint64{},
-			wantErr: false,
-		},
-		{
-			name: "Year 3, Subset 0, shard committee size = 61",
-			fields: fields{
-				beaconCommitteeStateSlashingBase: beaconCommitteeStateSlashingBase{
-					assignRule: NewAssignRuleV2(),
-					beaconCommitteeStateBase: beaconCommitteeStateBase{
-						beaconCommittee: []string{
-							key0, key, key2, key3, key, key2, key3,
-						},
-						shardCommittee: map[byte][]string{
-							0: []string{},
-							1: []string{
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0,
-							},
-							2: []string{},
-							3: []string{},
-							4: []string{},
-							5: []string{},
-							6: []string{},
-							7: []string{},
-						},
-						mu: &sync.RWMutex{},
-					},
-				},
-			},
-			args: args{
-				env: &SplitRewardEnvironment{
-					DAOPercent:                8,
-					ShardID:                   1,
-					IsSplitRewardForCustodian: false,
-					PercentCustodianReward:    0,
-					TotalReward: map[common.Hash]uint64{
-						common.PRVCoinID: 452696,
-					},
-					MaxSubsetCommittees: 2,
-					SubsetID:            0,
-				},
-			},
-			want: map[common.Hash]uint64{
-				common.PRVCoinID: 11433,
-			},
-			want1: map[common.Hash]uint64{
-				common.PRVCoinID: 405048,
-			},
-			want2: map[common.Hash]uint64{
-				common.PRVCoinID: 36215,
-			},
-			want3:   map[common.Hash]uint64{},
-			wantErr: false,
-		},
-		{
-			name: "Year 3, Subset 1, shard committee size = 61",
-			fields: fields{
-				beaconCommitteeStateSlashingBase: beaconCommitteeStateSlashingBase{
-					assignRule: NewAssignRuleV2(),
-					beaconCommitteeStateBase: beaconCommitteeStateBase{
-						beaconCommittee: []string{
-							key0, key, key2, key3, key, key2, key3,
-						},
-						shardCommittee: map[byte][]string{
-							0: []string{},
-							1: []string{
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0, key, key2, key3, key4, key5,
-								key0,
-							},
-							2: []string{},
-							3: []string{},
-							4: []string{},
-							5: []string{},
-							6: []string{},
-							7: []string{},
-						},
-						mu: &sync.RWMutex{},
-					},
-				},
-			},
-			args: args{
-				env: &SplitRewardEnvironment{
-					DAOPercent:                8,
-					ShardID:                   1,
-					IsSplitRewardForCustodian: false,
-					PercentCustodianReward:    0,
-					TotalReward: map[common.Hash]uint64{
-						common.PRVCoinID: 452696,
-					},
-					MaxSubsetCommittees: 2,
-					SubsetID:            1,
-				},
-			},
-			want: map[common.Hash]uint64{
-				common.PRVCoinID: 11804,
-			},
-			want1: map[common.Hash]uint64{
-				common.PRVCoinID: 404677,
-			},
-			want2: map[common.Hash]uint64{
-				common.PRVCoinID: 36215,
-			},
-			want3:   map[common.Hash]uint64{},
-			wantErr: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			b := &BeaconCommitteeStateV3{
-				beaconCommitteeStateSlashingBase: tt.fields.beaconCommitteeStateSlashingBase,
-				syncPool:                         tt.fields.syncPool,
-			}
-			got, got1, got2, got3, err := b.SplitReward(tt.args.env)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("BeaconCommitteeStateV3.SplitReward() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("BeaconCommitteeStateV3.SplitReward() got = %v, want %v", got, tt.want)
-			}
-			if !reflect.DeepEqual(got1, tt.want1) {
-				t.Errorf("BeaconCommitteeStateV3.SplitReward() got1 = %v, want %v", got1, tt.want1)
-			}
-			if !reflect.DeepEqual(got2, tt.want2) {
-				t.Errorf("BeaconCommitteeStateV3.SplitReward() got2 = %v, want %v", got2, tt.want2)
-			}
-			if !reflect.DeepEqual(got3, tt.want3) {
-				t.Errorf("BeaconCommitteeStateV3.SplitReward() got3 = %v, want %v", got3, tt.want3)
-			}
-		})
-	}
-}
