+++ dataaccessobject/statedb/stateobject_staker.go
@@ -10,35 +10,35 @@ import (
 )
 
 //@NOTE this struct is view object only
-type StakerInfoV2 struct {
+type StakerInfoSlashingVersion struct {
 	committeePublicKey string
 	rewardReceiver     key.PaymentAddress
 	txStakingID        common.Hash
 	autoStaking        bool
 }
 
-func NewStakerInfoV2WithCommittee(committeePublicKey string) *StakerInfoV2 {
-	return &StakerInfoV2{committeePublicKey: committeePublicKey}
+func NewStakerInfoSlashingVersionWithCommittee(committeePublicKey string) *StakerInfoSlashingVersion {
+	return &StakerInfoSlashingVersion{committeePublicKey: committeePublicKey}
 }
 
-func (s StakerInfoV2) CommitteePublicKey() string {
+func (s StakerInfoSlashingVersion) CommitteePublicKey() string {
 	return s.committeePublicKey
 }
 
-func (s StakerInfoV2) RewardReceiver() key.PaymentAddress {
+func (s StakerInfoSlashingVersion) RewardReceiver() key.PaymentAddress {
 	return s.rewardReceiver
 }
 
-func (s StakerInfoV2) TxStakingID() common.Hash {
+func (s StakerInfoSlashingVersion) TxStakingID() common.Hash {
 	return s.txStakingID
 }
 
-func (s StakerInfoV2) AutoStaking() bool {
+func (s StakerInfoSlashingVersion) AutoStaking() bool {
 	return s.autoStaking
 }
 
-func NewStakerInfoV2(committeePublicKey string, s *StakerInfo) *StakerInfoV2 {
-	return &StakerInfoV2{
+func NewStakerInfoSlashingVersion(committeePublicKey string, s *StakerInfo) *StakerInfoSlashingVersion {
+	return &StakerInfoSlashingVersion{
 		committeePublicKey: committeePublicKey,
 		rewardReceiver:     s.rewardReceiver,
 		txStakingID:        s.txStakingID,
